+++ samples/ModelView/Model.h
@@ -99,7 +99,7 @@ public:
 
     // implementation of ItemModel
     virtual bool isValidIndex(const CEGUI::ModelIndex& model_index) const;
-    virtual CEGUI::ModelIndex makeIndex(int child, const CEGUI::ModelIndex& parent_index);
+    virtual CEGUI::ModelIndex makeIndex(size_t child, const CEGUI::ModelIndex& parent_index);
     virtual CEGUI::ModelIndex getParentIndex(const CEGUI::ModelIndex& model_index);
     virtual int getChildCount(const CEGUI::ModelIndex& model_index);
     virtual CEGUI::String getData(const CEGUI::ModelIndex& model_index, CEGUI::ItemDataRole role = CEGUI::IDR_Text);
